se modifica slice de preconsulta
