+++ src/app/Dashboard/Inicio_Paciente/Preconsulta/[id]/page.jsx
@@ -3,7 +3,7 @@
 import Elboton from "@/components/Buttons/Elboton";
 import InputDiagnostico from "@/components/consulta/inputDiagnostico";
 import InputConsulta from "@/components/consulta/inputconsulta";
-import InputFile from "@/components/consulta/inputfile";
+import InputFile from "@/components/preconsulta/estudios";
 import IconRegresar from "@/components/icons/iconRegresar";
 import PreconsultaQuestion from "@/components/preconsulta/PreconsultaQuestion";
 import Link from "next/link";
