+++ src/components/preconsulta/estudios.jsx
@@ -3,8 +3,11 @@
 import Image from "next/image";
 import circleData from "@/components/images/circleData.png";
 import FileUpload from "@/components/consulta/file";
+import { useAppSelector } from "@/redux/hooks";
 
 export default function InputFile({ title, defaultOpen = false }) {
+    const formData = useAppSelector((state) => state.preconsultaForm.formData);
+
     return (
         <div className="flex flex-col">
             <details open={defaultOpen}>
@@ -16,17 +19,9 @@ export default function InputFile({ title, defaultOpen = false }) {
                         </p>
                     </div>
                 </summary>
-                <FileUpload label={"Electrocardiograma"} />
-                <FileUpload label={"RX de Torax"} />
-                <FileUpload label={"Ecocardiograma"} />
-                <FileUpload label={"Test de caminata"} />
-                <FileUpload label={"Funcional respiratorio"} />
-                <FileUpload label={"Tomografías"} />
-                <FileUpload label={"Cateterismo cardiaco derecho"} />
-                <FileUpload label={"CCG (Coronariografia)"} />
-                <FileUpload label={"Resonancia"} />
-                <FileUpload label={"Cateterismo cardiaco izquierdo"} />
-                <FileUpload label={"Otros estudios"} />
+                {Object.entries(formData.estudios).map(([key, value]) => (
+                    <FileUpload label={value.title} />
+                ))}
             </details>
         </div>
     );
